Refactor watchSessionConnection method via InstanceContextAware
